chore: refactor main.go
